[Refactor]: change deratory structure
