+++ app/api/v1/character_router.py
@@ -6,7 +6,7 @@ from app.database import get_db
 from app.lib import jwt_util
 from app.schemas.character_schema import CharacterCreate, CharacterResponse, CharacterUpdate
 
-from . import character_service
+from ...services import character_service
 
 router = APIRouter(
     prefix="/api/characters",
