+++ app/crud/auth_crud.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import Session
 
-from ...models.user import User
-from ...schemas.user_schema import UserCreate
+from ..models.user import User
+from ..schemas.user_schema import UserCreate
 
 
 def create_user(db: Session, user: UserCreate):
